requirements.txt
Django==1.6
django-tastypie==0.12.0
gnureadline==6.3.3
ipython==2.2.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.8.0
wsgiref==0.1.2
